Reduce image resolution if too large

+++ src/image-to-xlsx/page.py
@@ -4,12 +4,14 @@ import pickle
 import io
 import boto3
 import time
+from definitions import MAX_TEXTRACT_SYNC_SIZE
 from table import Table
 from unskewing import correct_skew
 from binarization import binarize
 from PIL import Image
 from surya.detection import batch_text_detection
 from surya.layout import batch_layout_detection
+from utils import maybe_reduce_resolution
 
 
 class Page:
@@ -163,64 +165,55 @@ class Page:
         return self.build_textract_tables_from_response(response)
 
     def get_textract_response(self):
+        self.page = maybe_reduce_resolution(self.page)
+        self.page.show()
         textract = boto3.client("textract")
-        # Upload file directly (small enough)
-        try:
-            img_byte_arr = io.BytesIO()
-            self.page.save(img_byte_arr, format="PNG")
-            img_byte_arr = img_byte_arr.getvalue()
+        img_byte_arr = io.BytesIO()
+        self.page.save(img_byte_arr, format="PNG")
+        img_byte_arr = img_byte_arr.getvalue()
 
+        # Analyze file directly (small enough)
+        if len(img_byte_arr) <= MAX_TEXTRACT_SYNC_SIZE:
             return textract.analyze_document(
                 Document={"Bytes": img_byte_arr},
                 FeatureTypes=["TABLES"],
             )
         # Too large, first upload to S3 and run asynchronous textract
-        except:
-            print("imhereeeeee")
-
+        else:
             # Upload the document to S3
-            bucket_name = 'test-textract-large-files'
+            bucket_name = "test-textract-large-files"
             file_name = "tmp_extract_page"
-            s3 = boto3.client('s3')
+            s3 = boto3.client("s3")
             # s3.upload_file(file_name, bucket_name, file_name)
-            s3.put_object(Bucket=bucket_name, Key=file_name, Body=self.page)
+            s3.put_object(Bucket=bucket_name, Key=file_name, Body=img_byte_arr)
 
             # Call Textract to analyze the document
             response = textract.start_document_analysis(
-                DocumentLocation={'S3Object': {'Bucket': bucket_name, 'Name': file_name}},
+                DocumentLocation={
+                    "S3Object": {"Bucket": bucket_name, "Name": file_name}
+                },
                 FeatureTypes=["TABLES"],  # Extract tables and forms; omit for raw text
             )
-            
-            print("my aws response", response)
-            
-            job_id = response['JobId']
-            print(f"Started Document Analysis Job. Job ID: {job_id}")
-            
+
+            job_id = response["JobId"]
+
             # Poll for job completion
             while True:
                 response = textract.get_document_analysis(JobId=job_id)
-                status = response['JobStatus']
-                print(f"Job Status: {status}")
-                if status in ['SUCCEEDED', 'FAILED']:
+                status = response["JobStatus"]
+                if status in ["SUCCEEDED", "FAILED"]:
                     break
                 time.sleep(5)
-            
-            if status == 'SUCCEEDED':
-                with open("test_cust.pkl", "wb") as f:
-                    pickle.dump(response, f)
-            else:
-                print("Document analysis failed.")
-            
+
             # Delete the file from S3 after processing
             s3.delete_object(Bucket=bucket_name, Key=file_name)
             print(f"Deleted {file_name} from bucket {bucket_name}")
 
+            return response
+
     def build_textract_tables_from_response(self, response):
         blocks = response["Blocks"]
-        print("response", response)
-
         file_tables = [block for block in blocks if block["BlockType"] == "TABLE"]
-        print("tables", file_tables)
         id_to_block = {block["Id"]: block for block in blocks}
 
         tables = []
